+++ alpha-web/src/main/java/com/alpha/web/module/screen/user/UserInvalid.java
@@ -18,7 +18,7 @@ public class UserInvalid extends BaseAjaxModule {
     @Resource
     private SystemAccountManager systemAccountManager;
 
-    public void execute(@Param("id") Long id, Context context) {
+    public void execute(@Param("id") int id, Context context) {
         Result<String> result = new Result<String>();
         try {
             SystemAccountDO systemAccountDO = new SystemAccountDO();
