+++ alpha-web/src/main/java/com/alpha/web/module/screen/vehicle/VehicleInvalid.java
@@ -20,7 +20,7 @@ public class VehicleInvalid extends BaseAjaxModule {
     @Resource
     private VehicleManager vehicleManager;
 
-    public void execute(@Param("id") Long id,  Context context) {
+    public void execute(@Param("id") int id,  Context context) {
         Result<String> result = new Result<String>();
         try {
             VehicleDO vehicleDO = new VehicleDO();
